[sui-verifier] Move helper functions into more foundational crates
`sui-verifier` is part of the versioned execution layer but these
helpers work on types that should remain fixed across execution
versions, so move them into crates that are not intended to be
versioned.

Test Plan:

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```
