+++ crates/sui-adapter/src/adapter.rs
@@ -14,6 +14,7 @@ use move_binary_format::{
     binary_views::BinaryIndexedView,
     file_format::{AbilitySet, CompiledModule, LocalIndex, SignatureToken, StructHandleIndex},
 };
+use move_bytecode_utils::{format_signature_token, resolve_struct};
 use move_bytecode_verifier::{verify_module_with_config, VerifierConfig};
 use move_core_types::{
     account_address::AccountAddress,
@@ -382,7 +383,7 @@ fn additional_validation_layout(
         SignatureToken::Vector(inner) => additional_validation_layout(view, inner)
             .map(|layout| PrimitiveArgumentLayout::Vector(Box::new(layout))),
         SignatureToken::StructInstantiation(idx, targs) => {
-            let resolved_struct = sui_verifier::resolve_struct(view, *idx);
+            let resolved_struct = resolve_struct(view, *idx);
             if resolved_struct == RESOLVED_STD_OPTION && targs.len() == 1 {
                 additional_validation_layout(view, &targs[0])
                     .map(|layout| PrimitiveArgumentLayout::Option(Box::new(layout)))
@@ -391,7 +392,7 @@ fn additional_validation_layout(
             }
         }
         SignatureToken::Struct(idx) => {
-            let resolved_struct = sui_verifier::resolve_struct(view, *idx);
+            let resolved_struct = resolve_struct(view, *idx);
             if resolved_struct == RESOLVED_SUI_ID {
                 Some(PrimitiveArgumentLayout::Address)
             } else if resolved_struct == RESOLVED_ASCII_STR {
@@ -596,7 +597,7 @@ fn type_check_struct(
     if !move_type_equals_sig_token(view, function_type_arguments, arg_type, param_type) {
         anyhow::bail!(
             "Expected argument of type {}, but found type {}",
-            sui_verifier::format_signature_token(view, param_type),
+            format_signature_token(view, param_type),
             arg_type
         )
     }
@@ -687,7 +688,7 @@ fn move_type_equals_struct_inst(
     param_type: StructHandleIndex,
     param_type_arguments: &[SignatureToken],
 ) -> bool {
-    let (address, module_name, struct_name) = sui_verifier::resolve_struct(view, param_type);
+    let (address, module_name, struct_name) = resolve_struct(view, param_type);
     let arg_type_params = arg_type.type_params();
 
     // same address, module, name, and type parameters
@@ -715,7 +716,7 @@ fn struct_tag_equals_struct_inst(
     param_type: StructHandleIndex,
     param_type_arguments: &[SignatureToken],
 ) -> bool {
-    let (address, module_name, struct_name) = sui_verifier::resolve_struct(view, param_type);
+    let (address, module_name, struct_name) = resolve_struct(view, param_type);
 
     // same address, module, name, and type parameters
     &arg_type.address == address
