+++ crates/sui-json/Cargo.toml
@@ -20,6 +20,7 @@ sui-types = { path = "../sui-types" }
 sui-verifier = { path = "../sui-verifier" }
 
 move-binary-format.workspace = true
+move-bytecode-utils.workspace = true
 move-core-types.workspace = true
 fastcrypto = { workspace = true }
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
