+++ crates/sui-json/src/lib.rs
@@ -11,6 +11,7 @@ use move_binary_format::{
     access::ModuleAccess, binary_views::BinaryIndexedView, file_format::SignatureToken,
     file_format_common::VERSION_MAX,
 };
+use move_bytecode_utils::resolve_struct;
 use move_core_types::account_address::AccountAddress;
 use move_core_types::identifier::IdentStr;
 use move_core_types::u256::U256;
@@ -622,7 +623,7 @@ pub fn primitive_type(
             }
         }
         SignatureToken::Struct(struct_handle_idx) => {
-            let resolved_struct = sui_verifier::resolve_struct(view, *struct_handle_idx);
+            let resolved_struct = resolve_struct(view, *struct_handle_idx);
             if resolved_struct == RESOLVED_ASCII_STR {
                 (
                     true,
@@ -662,7 +663,7 @@ pub fn primitive_type(
             }
         }
         SignatureToken::StructInstantiation(idx, targs) => {
-            let resolved_struct = sui_verifier::resolve_struct(view, *idx);
+            let resolved_struct = resolve_struct(view, *idx);
             // is option of a primitive
             if resolved_struct == RESOLVED_STD_OPTION && targs.len() == 1 {
                 // there is no MoveLayout for this so while we can still report whether a type
