+++ crates/sui-verifier/Cargo.toml
@@ -9,6 +9,7 @@ publish = false
 
 [dependencies]
 move-binary-format.workspace = true
+move-bytecode-utils.workspace = true
 move-bytecode-verifier.workspace = true
 move-core-types.workspace = true
